Feature/cierra-alentar-movimientos-enemigos

+++ Assets/Scripts/Personaje/Vida_Pj.cs
@@ -3,15 +3,15 @@ using UnityEngine;
 public class Vida_Pj : MonoBehaviour
 {
     // script simple para manejar la vida, el daño se manejará desde los enemigos :)
-    [SerializeField] private int vidaMaxima = 3;
-    private int vidaActual;
+    [SerializeField] private float vidaMaxima = 3;
+    private float vidaActual;
     void Start()
     {
         vidaActual = vidaMaxima;
     }
 
 
-    public void TomarDanio(int danio)
+    public void TomarDanio(float danio)
     {
         vidaActual -= danio;
         if (vidaActual <= 0)
